docs: show timer race interrupt composition
