+++ samples/invoke-process-timer-race-agent/langgraph.json
@@ -1,7 +1,7 @@
 {
   "dependencies": ["."],
   "graphs": {
-    "timeout-parent-agent": "./graph.py:parent_graph"
+    "timer-race-parent-agent": "./graph.py:parent_graph"
   },
   "env": ".env"
 }
